Revert "Fix audio/video slipping"
This reverts commit fe1f7067d672db61fb7e04d3a3f44d1404ba0b7c.

+++ libobs/obs-internal.h
@@ -326,7 +326,7 @@ struct obs_core_audio {
 
 	uint64_t buffered_ts;
 	struct circlebuf buffered_timestamps;
-        uint64_t buffering_wait_ticks;
+	int buffering_wait_ticks;
 	int total_buffering_ticks;
 
 	float user_volume;
